+++ report/schemas.api.md
@@ -1837,8 +1837,8 @@ export namespace World {
 // src/dapps/sale.ts:18:3 - (ae-incompatible-release-tags) The symbol "network" is marked as @public, but its signature references "Network" which is marked as @alpha
 // src/dapps/sale.ts:19:3 - (ae-incompatible-release-tags) The symbol "chainId" is marked as @public, but its signature references "ChainId" which is marked as @alpha
 // src/dapps/sale.ts:42:3 - (ae-incompatible-release-tags) The symbol "network" is marked as @public, but its signature references "Network" which is marked as @alpha
-// src/platform/item/emote/adr95/emote-data-adr95.ts:11:3 - (ae-incompatible-release-tags) The symbol "representations" is marked as @public, but its signature references "EmoteRepresentationADR73" which is marked as @alpha
-// src/platform/item/emote/adr95/representation-adr95.ts:10:3 - (ae-forgotten-export) The symbol "BodyShape" needs to be exported by the entry point index.d.ts
+// src/platform/item/emote/adr73/emote-data-adr73.ts:11:3 - (ae-incompatible-release-tags) The symbol "representations" is marked as @public, but its signature references "EmoteRepresentationADR73" which is marked as @alpha
+// src/platform/item/emote/adr73/representation-adr73.ts:10:3 - (ae-forgotten-export) The symbol "BodyShape" needs to be exported by the entry point index.d.ts
 // src/platform/item/third-party-props.ts:10:3 - (ae-incompatible-release-tags) The symbol "merkleProof" is marked as @public, but its signature references "MerkleProof" which is marked as @alpha
 // src/platform/scene/spawn-point.ts:10:3 - (ae-forgotten-export) The symbol "SinglePosition" needs to be exported by the entry point index.d.ts
 // src/platform/scene/spawn-point.ts:10:3 - (ae-forgotten-export) The symbol "MultiPosition" needs to be exported by the entry point index.d.ts
