+++ src/platform/item/emote/adr73/emote-data-adr73.ts
@@ -4,7 +4,7 @@ import {
   JSONSchema,
   ValidateFunction
 } from '../../../../validation'
-import { EmoteRepresentationADR73 } from './representation-adr95'
+import { EmoteRepresentationADR73 } from './representation-adr73'
 
 export type EmoteDataADR73 = {
   category: EmoteCategory
