+++ src/platform/item/emote/adr73/representation-adr73.ts
