+++ src/platform/item/emote/index.ts
@@ -1,3 +1,3 @@
 export { Emote } from './emote'
-export { EmoteRepresentationADR73 } from './adr95/representation-adr95'
-export { EmoteDataADR73 } from './adr95/emote-data-adr95'
+export { EmoteRepresentationADR73 } from './adr73/representation-adr73'
+export { EmoteDataADR73 } from './adr73/emote-data-adr73'
